requirements.txt
python-docx==0.8.11
python-dotenv==0.20.0
python-telegram-bot==13.12
pymongo[srv]==4.1.1